Fixed bug

+++ Travel App/Detail/Model/DetailModel.swift
@@ -44,8 +44,9 @@ class DetailModel {
         request.predicate = NSPredicate(format: "bookmarkTitle CONTAINS[cd] %@", name)
         do {
             let itemArray = try context.fetch(request)
-            print(itemArray.count)
-            context.delete(itemArray.first!)
+            if !itemArray.isEmpty {
+                context.delete(itemArray.first!)
+            }
         } catch {
             print("Error fetching data from context, \(error)")
         }
